Gave Collid_Handler Functionality
Given Functionality to the Collid_Handler with Blank sections for later use as I was not absolutely sure what to make the Handler do in this scenario or the Functionality currently does not exist. the Function is to allow Tags to be in a Blacklist and Whitelist, causing specific Functionality for each. additionally applied the function to the Player, Enemy, Bullet, and Grenade.

Added Collision to Bullet and Grenade that will most likely need to be modified later.