Made changes to bullet script and merged target script with health for handling of damage and death events
Bullet now checks to see what tag the object that is being hit has and ignores the player so it can't accidentally deal damage. Health script now handles taking damage as well as healing to be implemented with a pickup later on. Also added rigidbody to enemy so that physics events with bullets can take place. Using T key to test player health bar functionality