Added a Main Menu With a Start and Exit Game
Added a Main Menu which for testing purposes and to make it clear that it was working (as functionality is only present in Builds of the game), I have added Debug logs to them.

this took me so long because I did not realise it actually has to be attached to the Main Camera to work, and since I did not know this, it took me far longer than I thought.